move scheduling functions to separate file
